fix(arch): resolve check-arch.sh issues and usize violations
Script fixes:
- Fix Layer 4 firewall check to ignore comments (was matching "# aura-simulator")
- Fix builder method check unbound variable when no builders found
- Add aura-macros to runtime coupling allowlist (test utilities)

Style fixes (usize → sized types):
- SyncStats: facts_sent/received → u64
- AgentState: recovery_approvals/messages_sent → u32
- AgentStatus: recovery_approvals/messages_sent → u32

+++ crates/aura-sync/src/protocols/namespaced_sync.rs
@@ -87,9 +87,9 @@ pub struct SyncResponse {
 #[derive(Debug, Clone, Default)]
 pub struct SyncStats {
     /// Facts sent
-    pub facts_sent: usize,
+    pub facts_sent: u64,
     /// Facts received
-    pub facts_received: usize,
+    pub facts_received: u64,
     /// Sync duration in ms
     pub duration_ms: u64,
 }
